Rename scheduler modules to use snake case
