+++ src/smif/controller/job/__init__.py
@@ -3,8 +3,8 @@ Provides all the job run schedulers.
 """
 
 # import classes for access like ::
-#         from smif.controller.run import SerialJobScheduler
-from smif.controller.job.serialjobscheduler import SerialJobScheduler
+#         from smif.controller.job import SerialJobScheduler
+from smif.controller.job.serial_job_scheduler import SerialJobScheduler
 
 # Define what should be imported as * ::
 #         from smif.controller.job import *
