+++ src/smif/controller/modelrun.py
@@ -20,7 +20,7 @@ ModeRun has attributes:
 from logging import getLogger
 
 import networkx as nx
-from smif.controller.job.serialjobscheduler import SerialJobScheduler
+from smif.controller.job import SerialJobScheduler
 from smif.decision.decision import DecisionManager
 from smif.exception import SmifModelRunError, SmifTimestepResolutionError
 from smif.metadata import RelativeTimestep
