+++ src/smif/controller/run/__init__.py
@@ -2,10 +2,10 @@
 Provides all the model run schedulers.
 """
 
+from smif.controller.run.dafni_run_scheduler import DAFNIRunScheduler
 # import classes for access like ::
 #         from smif.controller.run import SubProcessRunScheduler
-from smif.controller.run.defaultscheduler import SubProcessRunScheduler
-from smif.controller.run.dafnirunscheduler import DAFNIRunScheduler
+from smif.controller.run.subprocess_run_scheduler import SubProcessRunScheduler
 
 # Define what should be imported as * ::
 #         from smif.controller.run import *
